fix React import usage
