+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as CourseEditor } from './CourseEditor';
+export { default as CourseSelector } from './CourseSelector';
+export { default as PlayerSetup } from './PlayerSetup';
+export { default as ScoreCard } from './ScoreCard';
